requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.3
paystack==2.2.0